pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "video-studio"
version = "0.1.0"
description = "Pipeline de edição de vídeo com IA — ChatCoreLab"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "python-dotenv",
    "pillow",
    "numpy",
]

[project.optional-dependencies]
manim = ["manim"]

[tool.setuptools.packages.find]
where = ["."]
